Hand-merge refactoring of finalized_page_cache_job.*
